+++ src/main/java/com/github/correa/finalreality/model/party/PlayerParty.java
@@ -1,28 +1,35 @@
-package com.github.correa.finalreality.model.Party;
+package com.github.correa.finalreality.model.party;
 
-import com.github.correa.finalreality.model.character.player.PlayerCharacter;
+import com.github.correa.finalreality.model.character.player.AbstractPlayerCharacter;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
+/**
+ * A class that has all the player characters in the party.
+ *
+ * @author Ignacio Slater Muñoz.
+ * @author Benjamín Correa Karstulovic.
+ */
+
 public class PlayerParty {
 
-    private final List<PlayerCharacter> playerParty = new ArrayList<>();
+    private final List<AbstractPlayerCharacter> playerParty = new ArrayList<>();
 
     /**
-     *Adds a PlayerCharacter to the PlayerParty
+     *Adds a PlayerCharacter to the PlayerParty.
      */
-    public void addPlayerCharacter(PlayerCharacter playerCharacter) {
+    public void addPlayerCharacter(AbstractPlayerCharacter playerCharacter) {
         if (playerParty.size() < 4){
             playerParty.add(playerCharacter);
         }
     }
 
     /**
-     * Return's the actual PlayerParty
+     * Return's the actual PlayerParty.
      */
-    public List<PlayerCharacter> getPlayerParty() {
+    public List<AbstractPlayerCharacter> getPlayerParty() {
         return List.copyOf(playerParty);
     }