+++ src/main/java/com/github/correa/finalreality/model/spell/SpellType.java
@@ -1,0 +1,11 @@
+package com.github.correa.finalreality.model.spell;
+
+/**
+ * Enumeration of all the spell types.
+ *
+ * @author Ignacio Slater Muñoz.
+ * @author Benjamín Correa Karstulovic.
+ */
+public enum SpellType {
+  BLACK, WHITE
+}