+++ src/test/java/com/github/cc3002/finalreality/model/inventory/WeaponWeaponInventoryTest.java
@@ -1,6 +1,6 @@
-package com.github.cc3002.finalreality.model.Inventory;
+package com.github.cc3002.finalreality.model.inventory;
 
-import com.github.correa.finalreality.model.Inventory.WeaponInventory;
+import com.github.correa.finalreality.model.inventory.WeaponInventory;
 import com.github.correa.finalreality.model.weapon.Weapon;
 import com.github.correa.finalreality.model.weapon.WeaponType;
 import org.junit.jupiter.api.BeforeEach;
@@ -8,14 +8,23 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+/**
+ * Set of tests for the WeaponInventory class.
+ *
+ * @author Ignacio Slater Muñoz.
+ * @author Benjamín Correa Karstulovic.
+ */
 public class WeaponWeaponInventoryTest {
   private WeaponInventory testWeaponInventory;
 
   @BeforeEach
   void setUp() { testWeaponInventory = new WeaponInventory(); }
 
+  /**
+   * Checks that the weapon inventory constructor and equals method works properly.
+   */
   @Test
-  void basicTest(){
+  void constructorTest(){
     var expectedWeaponInventory = new WeaponInventory();
     var sameWeaponInventory = testWeaponInventory;
     assertEquals(expectedWeaponInventory, testWeaponInventory);
@@ -24,6 +33,9 @@ public class WeaponWeaponInventoryTest {
     assertEquals(sameWeaponInventory, testWeaponInventory);
   }
 
+  /**
+   * Checks that the inventory works properly.
+   */
   @Test
   void InventoryTest(){
     assertTrue(testWeaponInventory.getWeaponInventory().isEmpty());