+++ src/test/java/com/github/cc3002/finalreality/model/party/EnemyPartyTest.java
@@ -1,6 +1,6 @@
-package com.github.cc3002.finalreality.model.Party;
+package com.github.cc3002.finalreality.model.party;
 
-import com.github.correa.finalreality.model.Party.EnemyParty;
+import com.github.correa.finalreality.model.party.EnemyParty;
 import com.github.correa.finalreality.model.character.Enemy;
 import com.github.correa.finalreality.model.character.ICharacter;
 import org.junit.jupiter.api.BeforeEach;
@@ -11,15 +11,26 @@ import java.util.concurrent.LinkedBlockingQueue;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+/**
+ * Set of tests for the EnemyParty class.
+ *
+ * @author Ignacio Slater Muñoz.
+ * @author Benjamín Correa Karstulovic.
+ */
 public class EnemyPartyTest {
+
   private EnemyParty testParty;
   protected BlockingQueue<ICharacter> turns;
   @BeforeEach
   void setUp() {
       testParty = new EnemyParty();
   }
+
+  /**
+   * Checks that the enemy party class constructor and equals method works properly.
+   */
   @Test
-  void basicTest() {
+  void constructorTest() {
     turns = new LinkedBlockingQueue<>();
     var expectedEnemyParty = new EnemyParty();
     var sameEnemyParty = testParty;
@@ -28,6 +39,10 @@ public class EnemyPartyTest {
     assertNotEquals(expectedEnemyParty, new Object());
     assertEquals(sameEnemyParty, testParty);
   }
+
+  /**
+   * Checks that the inventory works properly.
+   */
   @Test
   void EnemyPartySizeTest() {
     assertTrue(testParty.getEnemyParty().isEmpty());