+++ src/test/java/com/github/cc3002/finalreality/model/party/PlayerPartyTest.java
@@ -1,9 +1,8 @@
-package com.github.cc3002.finalreality.model.Party;
+package com.github.cc3002.finalreality.model.party;
 
-import com.github.correa.finalreality.model.Party.PlayerParty;
 import com.github.correa.finalreality.model.character.ICharacter;
-import com.github.correa.finalreality.model.character.player.CharacterClass;
-import com.github.correa.finalreality.model.character.player.PlayerCharacter;
+import com.github.correa.finalreality.model.character.player.classes.commonclasses.Engineer;
+import com.github.correa.finalreality.model.party.PlayerParty;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -12,6 +11,12 @@ import java.util.concurrent.LinkedBlockingQueue;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+/**
+ * Set of tests for the PlayerParty class.
+ *
+ * @author Ignacio Slater Muñoz.
+ * @author Benjamín Correa Karstulovic.
+ */
 public class PlayerPartyTest {
   private PlayerParty testParty;
   protected BlockingQueue<ICharacter> turns;
@@ -21,8 +26,11 @@ public class PlayerPartyTest {
     testParty = new PlayerParty();
   }
 
+  /**
+   * Checks that the player party class constructor and equals method works properly.
+   */
   @Test
-  void basicTest() {
+  void constructorTest() {
     turns = new LinkedBlockingQueue<>();
     var expectedPlayerParty = new PlayerParty();
     var samePlayerParty = testParty;
@@ -32,15 +40,18 @@ public class PlayerPartyTest {
     assertEquals(samePlayerParty, testParty);
   }
 
+  /**
+   * Checks that the inventory works properly.
+   */
   @Test
   void PlayerPartySizeTest() {
     assertTrue(testParty.getPlayerParty().isEmpty());
     for (int i = 0; i < 5; i++) {
       testParty.addPlayerCharacter(
-          new PlayerCharacter(turns, "Test 1", 10, 4, CharacterClass.ENGINEER));
+          new Engineer(turns, "Test 1", 10, 4));
     }
     testParty.addPlayerCharacter(
-        new PlayerCharacter(turns, "Test 1", 10, 4, CharacterClass.ENGINEER));
+        new Engineer(turns, "Test 1", 10, 4));
     assertEquals(4, testParty.getPlayerParty().size());
   }
 }